Restructured classes and added function to read events from a file in run mode
