--- modules/siddhi-launcher/src/main/java/org/wso2/siddhi/launcher/debug/util/DebugCallbackEvent.java
@@ -1,57 +1,0 @@
-/*
- * Copyright (c)  2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.siddhi.launcher.debug.util;
-
-import org.wso2.siddhi.core.debugger.SiddhiDebugger;
-import org.wso2.siddhi.core.event.ComplexEvent;
-
-/**
- * Event to be send in within the Debug callback
- */
-public class DebugCallbackEvent {
-    private ComplexEvent event;
-    private String queryName;
-    private int queryIndex;
-    private SiddhiDebugger.QueryTerminal queryTerminal;
-
-    public DebugCallbackEvent(String queryName, int queryIndex,
-                              SiddhiDebugger.QueryTerminal queryTerminal,
-                              ComplexEvent event) {
-        this.queryName = queryName;
-        this.queryIndex = queryIndex;
-        this.queryTerminal = queryTerminal;
-        this.event = event;
-    }
-
-    public String getQueryName() {
-        return queryName;
-    }
-
-    public int getQueryIndex() {
-        return queryIndex;
-    }
-
-    public ComplexEvent getEvent() {
-        return event;
-    }
-
-    public SiddhiDebugger.QueryTerminal getQueryTerminal() {
-        return queryTerminal;
-    }
-}
